+++ Tests/FluentAssertions.Specs/Collections/GenericCollectionAssertionsSpecs.cs
@@ -1284,6 +1284,192 @@ namespace FluentAssertions.Specs
 
         #endregion
 
+        #region Then be in order
+
+        [Fact]
+        public void When_the_collection_is_ordered_according_to_the_subsequent_ascending_assertion_it_should_succeed()
+        {
+            // Arrange
+            var collection = new[]
+            {
+                (1, "a"),
+                (2, "b"),
+                (2, "c"),
+                (3, "a")
+            };
+
+            // Act
+            Action action = () => collection.Should()
+                .BeInAscendingOrder(x => x.Item1)
+                .And
+                .ThenBeInAscendingOrder(x => x.Item2);
+
+            // Assert
+            action.Should().NotThrow();
+        }
+
+        [Fact]
+        public void When_the_collection_is_not_ordered_according_to_the_subsequent_ascending_assertion_it_should_fail()
+        {
+            // Arrange
+            var collection = new[]
+            {
+                (1, "a"),
+                (2, "b"),
+                (2, "c"),
+                (3, "a")
+            };
+
+            // Act
+            Action action = () => collection.Should()
+                .BeInAscendingOrder(x => x.Item1)
+                .And
+                .BeInAscendingOrder(x => x.Item2);
+
+            // Assert
+            action.Should().Throw<XunitException>()
+                .WithMessage("Expected collection * to be ordered \"by Item2\"*");
+        }
+
+        [Fact]
+        public void When_the_collection_is_ordered_according_to_the_subsequent_ascending_assertion_with_comparer_it_should_succeed()
+        {
+            // Arrange
+            var collection = new[]
+            {
+                (1, "a"),
+                (2, "B"),
+                (2, "b"),
+                (3, "a")
+            };
+
+            // Act
+            Action action = () => collection.Should()
+                .BeInAscendingOrder(x => x.Item1)
+                .And
+                .ThenBeInAscendingOrder(x => x.Item2, StringComparer.InvariantCultureIgnoreCase);
+
+            // Assert
+            action.Should().NotThrow();
+        }
+
+        [Fact]
+        public void When_the_collection_is_ordered_according_to_the_multiple_subsequent_ascending_assertions_it_should_succeed()
+        {
+            // Arrange
+            var collection = new[]
+            {
+                (1, "a", 1.1),
+                (2, "b", 1.2),
+                (2, "c", 1.3),
+                (3, "a", 1.1)
+            };
+
+            // Act
+            Action action = () => collection.Should()
+                .BeInAscendingOrder(x => x.Item1)
+                .And
+                .ThenBeInAscendingOrder(x => x.Item2)
+                .And
+                .ThenBeInAscendingOrder(x => x.Item3);
+
+            // Assert
+            action.Should().NotThrow();
+        }
+
+        [Fact]
+        public void When_the_collection_is_ordered_according_to_the_subsequent_descending_assertion_it_should_succeed()
+        {
+            // Arrange
+            var collection = new[]
+            {
+                (3, "a"),
+                (2, "c"),
+                (2, "b"),
+                (1, "a")
+            };
+
+            // Act
+            Action action = () => collection.Should()
+                .BeInDescendingOrder(x => x.Item1)
+                .And
+                .ThenBeInDescendingOrder(x => x.Item2);
+
+            // Assert
+            action.Should().NotThrow();
+        }
+
+        [Fact]
+        public void When_the_collection_is_not_ordered_according_to_the_subsequent_descending_assertion_it_should_fail()
+        {
+            // Arrange
+            var collection = new[]
+            {
+                (3, "a"),
+                (2, "c"),
+                (2, "b"),
+                (1, "a")
+            };
+
+            // Act
+            Action action = () => collection.Should()
+                .BeInDescendingOrder(x => x.Item1)
+                .And
+                .BeInDescendingOrder(x => x.Item2);
+
+            // Assert
+            action.Should().Throw<XunitException>()
+                .WithMessage("Expected collection * to be ordered \"by Item2\"*");
+        }
+
+        [Fact]
+        public void When_the_collection_is_ordered_according_to_the_subsequent_descending_assertion_with_comparer_it_should_succeed()
+        {
+            // Arrange
+            var collection = new[]
+            {
+                (3, "a"),
+                (2, "b"),
+                (2, "B"),
+                (1, "a")
+            };
+
+            // Act
+            Action action = () => collection.Should()
+                .BeInDescendingOrder(x => x.Item1)
+                .And
+                .ThenBeInDescendingOrder(x => x.Item2, StringComparer.InvariantCultureIgnoreCase);
+
+            // Assert
+            action.Should().NotThrow();
+        }
+
+        [Fact]
+        public void When_the_collection_is_ordered_according_to_the_multiple_subsequent_descending_assertions_it_should_succeed()
+        {
+            // Arrange
+            var collection = new[]
+            {
+                (3, "a", 1.1),
+                (2, "c", 1.3),
+                (2, "b", 1.2),
+                (1, "a", 1.1)
+            };
+
+            // Act
+            Action action = () => collection.Should()
+                .BeInDescendingOrder(x => x.Item1)
+                .And
+                .ThenBeInDescendingOrder(x => x.Item2)
+                .And
+                .ThenBeInDescendingOrder(x => x.Item3);
+
+            // Assert
+            action.Should().NotThrow();
+        }
+
+        #endregion
+
         private class SomeClass
         {
             public string Text { get; set; }